Refactor auth modal
